@@ -12,7 +12,7 @@
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>2.7.1</version>
+		<version>3.0.0-SNAPSHOT</version>
 		<relativePath />
 	</parent>
 
@@ -22,10 +22,9 @@
 		<module>spring-modulith-events</module>
 		<module>spring-modulith-test</module>
 		<module>spring-modulith-docs</module>
-		<module>spring-modulith-observability</module>
+		<!--<module>spring-modulith-observability</module>-->
 		<module>spring-modulith-moments</module>
 		<module>spring-modulith-starter-jpa</module>
-		<module>spring-modulith-starter-jpa-jakarta</module>
 		<module>spring-modulith-starter-test</module>
 	</modules>
 
@@ -110,6 +109,13 @@ limitations under the License.
 					<releases>
 						<enabled>false</enabled>
 					</releases>
+				</repository>
+				<repository>
+					<id>spring-milestone</id>
+					<url>https://repo.spring.io/milestone</url>
+					<snapshots>
+						<enabled>false</enabled>
+					</snapshots>
 				</repository>
 			</repositories>
 		</profile>
@@ -201,8 +207,8 @@ limitations under the License.
 				<groupId>org.apache.maven.plugins</groupId>
 				<artifactId>maven-compiler-plugin</artifactId>
 				<configuration>
-					<source>1.8</source>
-					<target>1.8</target>
+					<source>17</source>
+					<target>17</target>
 				</configuration>
 			</plugin>
 
