@@ -15,8 +15,9 @@
  */
 package example.events;
 
-import javax.persistence.EntityManager;
-import javax.persistence.EntityManagerFactory;
+import jakarta.persistence.EntityManager;
+import jakarta.persistence.EntityManagerFactory;
+
 import javax.sql.DataSource;
 
 import org.springframework.context.annotation.Bean;
