@@ -15,6 +15,9 @@
  */
 package org.springframework.modulith.events.jpa;
 
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Data;
@@ -23,10 +26,6 @@ import lombok.RequiredArgsConstructor;
 
 import java.time.Instant;
 import java.util.UUID;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Id;
 
 /**
  * @author Oliver Gierke
